server/requirements.txt
Flask==1.0.2
Flask_Cors==3.0.6
Flask-GraphQL==2.0.0
Flask-JWT-Extended==3.10.0
Flask-Migrate==2.2.1
Flask-OAuthlib==0.9.4
Flask_Redis==0.3.0
Flask-SQLAlchemy==2.3.1
GeoAlchemy2==0.5.0
graphene==2.1.3
graphene_sqlalchemy==2.1.0
psycopg2==2.7.5

# Development
black==18.6b4
pytest==3.6.2
pytest-flask==0.10.0
snapshottest==0.5.0
